requirements.txt
click
dataclasses-json
pymodbus
pyserial-asyncio